Add metrics comparison
